Refactor import statements to simplify helper module access
