+++ Resolute/cogs/arenas.py
@@ -6,8 +6,7 @@ from discord.ext import commands
 
 from Resolute.bot import G0T0Bot, G0T0Context
 from Resolute.constants import CHANNEL_BREAK
-from Resolute.helpers.autocomplete import get_arena_type_autocomplete
-from Resolute.helpers.general_helpers import confirm
+from Resolute.helpers import confirm
 from Resolute.models.categories import ArenaTier, ArenaType
 from Resolute.models.embeds.arenas import ArenaPhaseEmbed, ArenaStatusEmbed
 from Resolute.models.embeds.players import ArenaPostEmbed
@@ -130,6 +129,11 @@ class Arenas(commands.Cog):
 
         raise G0T0Error("Something went wrong")
 
+    async def arena_type_autocomplete(
+        self, ctx: discord.AutocompleteContext
+    ) -> list[str]:
+        return [f.value for f in self.bot.compendium.arena_type[0].values()] or []
+
     @arena_commands.command(
         name="claim", description="Opens an arena in this channel and sets you as host"
     )
@@ -139,7 +143,7 @@ class Arenas(commands.Cog):
         type: discord.Option(
             discord.SlashCommandOptionType(3),
             description="Arena Type",
-            autocomplete=get_arena_type_autocomplete,
+            autocomplete=arena_type_autocomplete,
             required=True,
             default="COMBAT",
         ),
