+++ Resolute/cogs/rooms.py
@@ -4,7 +4,7 @@ import discord
 from discord.ext import commands
 
 from Resolute.bot import G0T0Bot, G0T0Context
-from Resolute.helpers.general_helpers import is_admin
+from Resolute.helpers import is_admin
 from Resolute.models.objects.exceptions import G0T0Error
 from Resolute.models.views.rooms import RoomSettingsUI
