+++ Resolute/models/views/shatterpoint.py
@@ -4,7 +4,7 @@ from typing import Mapping
 import discord
 
 from Resolute.bot import G0T0Bot
-from Resolute.helpers.general_helpers import confirm
+from Resolute.helpers import confirm
 from Resolute.models.categories.categories import CodeConversion, Faction
 from Resolute.models.embeds import ErrorEmbed
 from Resolute.models.embeds.shatterpoint import ShatterpointEmbed, ShatterpointLogEmbed
